+++ cpp/src/db/insert/MemTableFile.cpp
@@ -87,8 +87,14 @@ Status MemTableFile::Serialize() {
     table_file_schema_.file_size_ = execution_engine_->PhysicalSize();
     table_file_schema_.row_count_ = execution_engine_->Count();
 
-    table_file_schema_.file_type_ = (size >= table_file_schema_.index_file_size_) ?
-                                    meta::TableFileSchema::TO_INDEX : meta::TableFileSchema::RAW;
+    //if index type isn't IDMAP, set file type to TO_INDEX if file size execeed index_file_size
+    //else set file type to RAW, no need to build index
+    if (table_file_schema_.engine_type_ != (int)EngineType::FAISS_IDMAP) {
+        table_file_schema_.file_type_ = (size >= table_file_schema_.index_file_size_) ?
+                                        meta::TableFileSchema::TO_INDEX : meta::TableFileSchema::RAW;
+    } else {
+        table_file_schema_.file_type_ = meta::TableFileSchema::RAW;
+    }
 
     auto status = meta_->UpdateTableFile(table_file_schema_);